important changes to ModelBackedView
now, ModelBackedView will use a `attributes` array property, which will
contain the properties that will be used on the form.
You can use `bindAttributesTo` property to specify the source (by
default, it will be the view’s model).
You can use `getFormData` to get an object with the attributes
specified for the form.
the `defineComputedPropertiesForAttributes` method will define the
attributes (and bindings to it’s source) on the fly when the view is
created